requirements.txt
Flask==3.1.2
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-WTF==1.1.2
python-dotenv==1.0.0
pyserial==3.5
qrcode[pil]
reportlab==4.0.4
openpyxl==3.1.2